+++ src/Systems/CHIP8/Cores/SCHIP_MODERN.hpp
@@ -14,17 +14,17 @@
 /*==================================================================*/
 
 class SCHIP_MODERN final : public Chip8_CoreInterface {
-	static constexpr u32 cTotalMemory{ KiB(4) };
+	static constexpr u64 cTotalMemory{ KiB(4) };
 	static constexpr u32 cSafezoneOOB{    32 };
 	static constexpr u32 cGameLoadPos{   512 };
 	static constexpr u32 cStartOffset{   512 };
 	static constexpr f32 cRefreshRate{ 60.0f };
 
-	static constexpr s32 cResSizeMult{     8 };
-	static constexpr s32 cScreenSizeX{    64 };
-	static constexpr s32 cScreenSizeY{    32 };
-	static constexpr s32 cInstSpeedHi{    45 };
-	static constexpr s32 cInstSpeedLo{    30 };
+	static constexpr s32 cResSizeMult{  8 };
+	static constexpr s32 cScreenSizeX{ 64 };
+	static constexpr s32 cScreenSizeY{ 32 };
+	static constexpr s32 cInstSpeedHi{ 45 };
+	static constexpr s32 cInstSpeedLo{ 30 };
 
 	static constexpr u32 cMaxDisplayW{ 128 };
 	static constexpr u32 cMaxDisplayH{  64 };
